requirements.txt
django
djangorestframework
pillow
psycopg2-binary
drf-yasg
python-decouple
djangorestframework-simplejwt

